backend/requirements.txt
# Django Core
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# Database
psycopg2-binary>=2.9.0
dj-database-url>=2.1.0
python-dotenv>=1.0.1

# Authentication & Security
djangorestframework-simplejwt>=5.2.0
django-environ>=0.11.0

# Storage (S3/MinIO)
boto3>=1.28.0
django-storages>=1.14.0

# Celery
celery>=5.3.0
redis>=5.0.0

# API Documentation
drf-spectacular>=0.26.0

# Filtering
django-filter>=23.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Production Server
gunicorn>=21.2.0

# Development (optional, can be in requirements-dev.txt)
# ipython>=8.0.0
# django-debug-toolbar>=4.2.0
